fix: Detail on single media upload
